+++ cpp07/ex02/src/main.cpp
@@ -38,7 +38,7 @@ int main()
    Array<int> firstArr(5);
    std::cout << "firstArr[] size: " << firstArr.size() << std::endl;
    std::cout << "Contents: " << std::endl;
-   for (unsigned int i = 0; i < firstArr.size(); i++)
+   for (size_t i = 0; i < firstArr.size(); i++)
    {
       firstArr[i] = (i + 1) * 100;
       std::cout << "[" << firstArr[i] << "]" << std::endl;
@@ -55,7 +55,7 @@ int main()
    Array<int> secondArr(5);
    std::cout << "secondArr[] size: " << firstArr.size() << std::endl;
    std::cout << "Contents: " << std::endl;
-   for (unsigned int i = 0; i < secondArr.size(); i++)
+   for (size_t i = 0; i < secondArr.size(); i++)
    {
       secondArr[i] = (i + 1);
       std::cout << "[" << secondArr[i] << "]" << std::endl;
@@ -65,7 +65,7 @@ int main()
    std::cout << "Post- copy assignment from firstArr[]: " << std::endl;
    secondArr = firstArr;
    std::cout << "Updated secondArr[] contents:" << std::endl;
-   for (unsigned int i = 0; i < secondArr.size(); i++)
+   for (size_t i = 0; i < secondArr.size(); i++)
    {
       std::cout << "[" << secondArr[i] << "]" << std::endl;
    }
@@ -75,7 +75,7 @@ int main()
    Array<int> thirdArr(secondArr);
    std::cout << "thirdArr[] size: " << thirdArr.size() << std::endl;
    std::cout << "Contents copied from secondArr[]: " << std::endl;
-   for (unsigned int i = 0; i < thirdArr.size(); i++)
+   for (size_t i = 0; i < thirdArr.size(); i++)
    {
       std::cout << "[" << thirdArr[i] << "]" << std::endl;
    }
@@ -85,14 +85,14 @@ int main()
    thirdArr[2] = 42;
    thirdArr[4] = 42;
    std::cout << "Printing thirdArr contents with modifications:" << std::endl;
-   for (unsigned int i = 0; i < thirdArr.size(); i++)
+   for (size_t i = 0; i < thirdArr.size(); i++)
    {
       std::cout << "[" << thirdArr[i] << "]" << std::endl;
    }
    std::cout << std::endl;
    std::cout << " ________ CHAR ARRAY TESTS ____________ " << std::endl;
    Array<char> charArr(5);
-   for (unsigned int i = 0; i < thirdArr.size(); i++)
+   for (size_t i = 0; i < thirdArr.size(); i++)
    {
       charArr[i] = i + 'a';
       std::cout << "[" << charArr[i] << "]" << std::endl;
